Check parameters
