+++ src/basic-execution.ts
@@ -4,62 +4,15 @@ import type {
   BatchStepModel,
   Ctx,
 } from './build-model.js';
-import { setDataValue } from './data-value-utils.js';
-import { getProperty } from 'dot-prop';
+import {
+  setDataValue,
+  getSupportedProperty,
+  isTruthy,
+  isFalsy,
+} from './data-value-utils.js';
 import { Result, succeed } from './railway.js';
 
 type BasicExecution = Result<Ctx, { message: string }>;
-
-const getDataProperty = (
-  valuePath: string,
-  value?: Record<string, AnyDataValue>
-): AnyDataValue | undefined => {
-  if (value === undefined) {
-    return undefined;
-  }
-  const [data_prefix, otherKeys] = valuePath.split('.');
-  if (
-    data_prefix === undefined ||
-    data_prefix !== 'data' ||
-    otherKeys === undefined
-  ) {
-    return undefined;
-  }
-
-  const [first, second, third] = otherKeys.split('::');
-  if (first === undefined || second === undefined || third === undefined) {
-    return undefined;
-  }
-  return value[`${first}::${second}::${third}`];
-};
-
-const getSupportedProperty = (
-  ctx: Ctx,
-  valuePath: string
-): AnyDataValue | undefined => {
-  const isInsideData = valuePath.startsWith('data.');
-  const value = isInsideData
-    ? getDataProperty(valuePath, ctx.data)
-    : getProperty(ctx, valuePath);
-  if (
-    typeof value === 'string' ||
-    typeof value === 'number' ||
-    typeof value === 'boolean' ||
-    typeof value === 'object'
-  ) {
-    return value;
-  } else return undefined;
-};
-
-const isFalsy = (value: unknown): boolean =>
-  value === false ||
-  value === undefined ||
-  value === null ||
-  value === 0 ||
-  (typeof value === 'string' && value.trim().length === 0) ||
-  (Array.isArray(value) && value.length === 0);
-
-const isTruthy = (value: unknown): boolean => !isFalsy(value);
 
 const getPropertyList = (
   ctx: Ctx,
